requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028898966/work
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer==2.0.12
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3
h5py==3.6.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numexpr @ file:///tmp/build/80754af9/numexpr_1640689833592/work
numpy==1.22.3
opencv-python==4.5.5.64
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.12.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-ignite==0.4.8
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1635187048948/work
scipy==1.8.0
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
snntorch @ file:///home/conda/feedstock_root/build_artifacts/snntorch_1646883320246/work
soupsieve==2.3.2.post1
stack-data==0.2.0
terminado==0.13.3
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tinycss2==1.1.1
torch @ https://download.pytorch.org/whl/cu113/torch-1.11.0%2Bcu113-cp39-cp39-linux_x86_64.whl
torchaudio==0.11.0+cu113
torchinfo==1.6.5
torchvision==0.12.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094708661/work
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
ujson==5.2.0
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
